+++ server/impl/src/main/java/com/walmartlabs/concord/server/project/ProjectConfigurationDao.java
@@ -2,49 +2,31 @@ package com.walmartlabs.concord.server.project;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.common.base.Throwables;
-import com.google.common.cache.CacheBuilder;
-import com.google.common.cache.CacheLoader;
-import com.google.common.cache.LoadingCache;
 import com.walmartlabs.concord.common.ConfigurationUtils;
 import com.walmartlabs.concord.common.db.AbstractDao;
 import org.jooq.Configuration;
 import org.jooq.DSLContext;
+import org.jooq.impl.DSL;
 
 import javax.inject.Inject;
 import javax.inject.Named;
 import javax.inject.Singleton;
-import java.io.ByteArrayInputStream;
-import java.io.ByteArrayOutputStream;
 import java.io.IOException;
-import java.io.InputStream;
-import java.time.Duration;
-import java.time.temporal.ChronoUnit;
 import java.util.List;
 import java.util.Map;
-import java.util.Optional;
-import java.util.concurrent.TimeUnit;
+
+import static com.walmartlabs.concord.server.jooq.public_.tables.Projects.PROJECTS;
 
 @Named
 @Singleton
 public class ProjectConfigurationDao extends AbstractDao {
 
-    public static final String PROJECT_CFG_ATTACHMENT_KEY = "_cfg";
-    private static final Duration CACHE_TTL = Duration.of(1, ChronoUnit.MINUTES);
-
-    private final ProjectAttachmentDao attachmentDao;
     private final ObjectMapper objectMapper;
 
-    // TODO #multiserver
-    private final LoadingCache<String, Optional<Map<String, Object>>> cache;
-
     @Inject
-    protected ProjectConfigurationDao(Configuration cfg, ProjectAttachmentDao attachmentDao) {
+    protected ProjectConfigurationDao(Configuration cfg) {
         super(cfg);
-        this.attachmentDao = attachmentDao;
         this.objectMapper = new ObjectMapper();
-        this.cache = CacheBuilder.newBuilder()
-                .expireAfterWrite(CACHE_TTL.getSeconds(), TimeUnit.SECONDS)
-                .build(new Loader());
     }
 
     public void insert(String projectName, Map<String, Object> cfg) {
@@ -52,21 +34,7 @@ public class ProjectConfigurationDao extends AbstractDao {
     }
 
     public void insert(DSLContext tx, String projectName, Map<String, Object> cfg) {
-        try (ByteArrayOutputStream out = new ByteArrayOutputStream()) {
-            objectMapper.writeValue(out, cfg);
-
-            InputStream in = new ByteArrayInputStream(out.toByteArray());
-            attachmentDao.insert(tx, projectName, PROJECT_CFG_ATTACHMENT_KEY, in);
-
-            cache.put(projectName, Optional.of(cfg));
-        } catch (IOException e) {
-            throw Throwables.propagate(e);
-        }
-    }
-
-    public void delete(DSLContext tx, String projectName) {
-        attachmentDao.delete(tx, projectName, PROJECT_CFG_ATTACHMENT_KEY);
-        cache.invalidate(projectName);
+        update(tx, projectName, cfg);
     }
 
     public void update(String projectName, Map<String, Object> value) {
@@ -74,21 +42,32 @@ public class ProjectConfigurationDao extends AbstractDao {
     }
 
     public void update(DSLContext tx, String projectName, Map<String, Object> cfg) {
-        delete(tx, projectName);
-        insert(tx, projectName, cfg);
-    }
+        byte[] ab;
+        try {
+            ab = objectMapper.writeValueAsBytes(cfg);
+        } catch (IOException e) {
+            throw Throwables.propagate(e);
+        }
 
-    public Map<String, Object> get(String projectName) {
-        return cache.getUnchecked(projectName).orElse(null);
+        tx.update(PROJECTS)
+                .set(PROJECTS.PROJECT_CFG, ab)
+                .where(PROJECTS.PROJECT_NAME.eq(projectName))
+                .execute();
     }
 
     @SuppressWarnings("unchecked")
-    private Map<String, Object> _get(String projectName) {
-        try (InputStream in = attachmentDao.get(projectName, PROJECT_CFG_ATTACHMENT_KEY)) {
-            if (in == null) {
+    public Map<String, Object> get(String projectName) {
+        try (DSLContext create = DSL.using(cfg)) {
+            byte[] ab = create.select(PROJECTS.PROJECT_CFG)
+                    .from(PROJECTS)
+                    .where(PROJECTS.PROJECT_NAME.eq(projectName))
+                    .fetchOne(PROJECTS.PROJECT_CFG);
+
+            if (ab == null) {
                 return null;
             }
-            return objectMapper.readValue(in, Map.class);
+
+            return objectMapper.readValue(ab, Map.class);
         } catch (IOException e) {
             throw Throwables.propagate(e);
         }
@@ -109,19 +88,5 @@ public class ProjectConfigurationDao extends AbstractDao {
             throw new IllegalArgumentException("Invalid data type, expected list, got: " + v.getClass());
         }
         return (List<Map<String, Object>>) v;
-    }
-
-    @SuppressWarnings("unchecked")
-    private class Loader extends CacheLoader<String, Optional<Map<String, Object>>> {
-
-        @Override
-        public Optional<Map<String, Object>> load(String key) throws Exception {
-            Map<String, Object> m = _get(key);
-            if (m == null) {
-                return Optional.empty();
-            }
-
-            return Optional.of(m);
-        }
     }
 }